Debug and test the 'ModPagespeedStatistics off' using the system-test.
Adds in a new testcase to explicitly validate a claim made on the discussion
list -- that we can rewrite CSS files across domains that have been mapped
together via ModPagespeedMapRewriteDomain.  Note that no functional changes were
needed, but I did a minor refactor in the test to pull together all the
'yellow' and 'blue' css text into char[] constants.

Fix a corner-case combining CSS files from different paths where a file
that changes the resolved base needs to be backed out cause it makes the
URL too big.  This was resulting in an extra level of hierarchy in the
encoded names, and could create functional issues on multi-server setups
where rewritten URL decoding is needed for functionality.

Make ScriptTagScanner understand various attributes that determine
whether something is JavaScript or not, whether it should run, and to
decode the various non-blocking execution modes.

Port existing filters to use ScriptTagScanner rather than ad-hoc JS
detection code that was different in all of and add testcases for
instances where the previous checks failed.