Change all references from [Simple]MetaData to ResponseHeaders & RequestHeaders.
Note that the parsing functionality is now split into a separate class, 
ResponseHeaderParser, that needs to be instantiated explicitly from 
any class that needs to parse HTTP streams. 

This does not switch to using protobufs for serialization yet - 
that's a cache-busting move that requires us to do something specific 
to ensure that we don't try to read the old-format files. 

Move more functionality from meta_data.cc to response_headers.cc and
request_headers.cc.

Get rid of the now-obsolete simple_meta_data[_test].(cc|h)