Rework ConditionalFetch interface.
More thorough tests for timer_based_abstract_lock.  The obvious properties were mostly being tested in subclasses, but Maks had run into issues with early / incorrect termination of the blocking methods.  Those are particularly tricky to test, and the code here is a bit delicate: we have to stand up a second thread, and have the main testing thread test that the second thread is blocking in the way we expect, then kill and clean up that second thread. (jmaessen)