Layer a mock thread system over a real thread system, so that we can control condition variables in mock time.
Two mechanisms are provided to work with mock time, though only one is used at
this point.

Method #1 is to use a new idle-callback provided by the Worker class.  Use that
to signal the condition-variable and allow the TimedWait in
RewriteDriver::Render to return.

Method #2, not in use yet, is to allow the various phases of RewriteContext and
MockUrlFetcher to advance time, and use new MockTimer alarms to signal the
condition-variable to complete the TimedWait as well.