Add gyp rule to compile json library (gagansingh)
Rework RewriteOptions::Merge to merge src into this, rather than one,two into
this.

This resolves a potential performance issue because of the way wildcards
are combined.  Because our basic mechanism for merging RewriteOptions
was dst=(one+two), we defined Clone in terms of CopyFrom, and CopyFrom
via Merge(one,one).  When 'one' has wildcards, they get doubled in
size.

The reason we used dst=one+two in the first place is that matches the .htaccess
semantics in Apache.  But it's easy to define that in terms of dst+=one,
dst+=two, so that's what this CL does, simplifying the Merge function.