Stop cache extending things that aren't CSS/JS/images. Before this CL, we would allow extending some things (content types we recognized), but only changing the Content-Type to text/plain.
So .ttf files would not be cache extended, because they were not a recognized content type, but .html (or .swf, etc.) files would be cache extended as .txt.

This came up because when CssFilter falls back to CssTagScanner, it no longer knows the context of URLs it is trying to rewrite, so it passes all to image rewrite filter and cache extender. We don't want to cache extend non-CSS/JS/images.