More uses of FALLTHROUGH_INTENDED. (sligocki)
Add HTTPS fetching support in Serf.  Note that the Serf support is
dependent on OpenSSL, which is large and has a complex license.  We do
not include it in our open-source distribution by default, and it must
be compiled form source after some small patches to source & config
files. (jmarantz)