factory-up-front: move driver factory instantiation up front
We need to have the NgxRewriteDriverFactory available at
configuration parse time to set options on it. This change moves
instantiation of the factory in front of configuration parsing.
Afterwards, a check is done to see if we where enabled anywhere.
If we where not enabled / do not have any active server contexts,
the NgxRewriteDriverFactory is deleted. It's not very elegant,
but we can't think of anything better currently.