statistics: fix statistics and add a test so we don't break it again
If we get a request for `/ngx_pagespeed_statistics` we don't create a
`ps_request_context`.  We create the content properly, but then in the header
filter we assume that "ctx is NULL iff this is not a resource fetch" and so
conclude that we're not a resource fetch.  Then we call `CreateRequestContext()`
with `resource_fetch=false` and so it doesn't check to see if it's statistics.

Here I fix this by pulling the check for statistics outside of the
resource-fetch-only part of `CreateRequestContext()`.