map-proxy-domain: support without tests
In pull request #247 I added tests for CDN support for MapProxyDomain, but this
depends on FetchProxy which depends on per-server-block fetchers.  In
mod_pagespeed there's a complicated map that means you have one serf fetcher for
each distinct value of FetchProxy, and we need that here too.

So for now, commit the minor supporting changes without the test code, and leave
the testing until we have the time to properly support FetchProxy.