Initial version for supporting caching of rewritten HTML in external caching layers (E.g. usecase: nginx's proxy_cache).
RewriteOptions have been added to indicate whether the feature is enabled or not, how much rewriting (before the response is served out) qualifies an entry for being retained in the cache, and the path to be used for cache purging. If the feature is enabled, if there is significant rewriting remaining to be done (after the response is served out), we detect this case in ReleaseRewriteDriver and then initiate a cache purge. The cache purge in the nginx case is a loopback request that hits the same server (which has a configured proxy_cache). 

Some obvious TODOs for future CLs: Support cache purging variations with other pagespeed-internal caches, and other external caching layers (varnish) in a generic manner. Fine tune the "rewritten-percentage" computation.