Only count HTTP cache expiration statistics to be incremented in the HTTPCache::Find() path.
Previously, this stat was being incremented in the Put() and "should we rewrite this file?" paths. Those cases are not really things expiring out of the cache so much as never ending up in the cache in the first place. This puts all the HTTP cache stats together and allows us to actually post an info message noting that specific URLs expired out of cache.