MapRewriteDomain shouldn't change Location headers in redirects.
In r3029 we changed ProxyFetch to use RewriteDomain mappings on Location
headers.  This was to support Adwords experiments that we're no longer
pursuing, and is not necessary any longer.  It also creates a bug in
ngx_pagespeed, where people don't expect MapRewriteDomain to affect redirects.
Reverting this back to how it was before.

Fixes https://github.com/pagespeed/ngx_pagespeed/issues/656