native-fetcher: keepalive handling changes
A few changes on top of @dinic his work on keep-alive support for the native fetcher.
Most notable changes:
- Request keep-alive by adding the appropriate request header
- Fixes connections getting reused while they are servicing other requests:
- Remove connection from the pool of available connections for keepalive when applicable
- Disable keepalive in more appropriate situations
- Response parsing fixes
- Remove connections that timeout from the k.a. pool
- Add a few sanity (D)CHECKS
- Emit debug messages for traceability

Note that this change does not reformat the indenting within the net_instaweb namespace
of ngx_fetch.cc, as that would make it very hard to assess what has changes from the diff.

Note: The 'PageSpeed CSS loaded in fallback mode is always chunked' tests must be disabled
when testing (currently fails on master without this change as well).