Changes by huibao: Use OptiPNG and multiple configurations to improve Scanline PNG writer
This CL improves the Scanline PNG writer by
(1) using OptiPNG
(2) trying 4 set of configurations and using the best one.

On 10,389 GIF and PNG images test set, the average size has been reduced from 8,551 to 8,186 bytes. That is 4.3% improvement.
-----

Change LOG level to INFO for images with size different from that of frames

This is a quick fix to silent the log. A thorough fix is to pad the frame when it's smaller than the image.