Add in-memory lock manager & test.
Note that this change does not incorporate the new lock manager into
any production code; it just adds a test.

To add this to a single-process server it needs a wrapper (or update)
to add thread safety.  To add this to MPS/NPS it needs a server/client
wrapper for serialization and an RPC infrastructure.

By itself this new lock manager does not provide a measurable benefit,
but it enables us to gauge popularity of a resource and use that to
avoid optimizing unpopular images.