jscompiler: use consistent ordering for the closure library files
Before this change the ordering of closure library files inside our
compiled js files would change depending on which computer the build
ran on.  This would lead to a lot of noise on commits, where unrelated
js 'changes' would get included in commits despite no actual changes
in the js source.

By explicitly telling closure compiler about all the js files in a
deterministic order the files it generate become deterministic as
well.  This is a bit of a hack, sticking a find | sort | sed
into the gyp file, but I think it's the least hacky way to fix it.