Threadsafe wrapper for MemLockManager.  This required a new API for Scheduler to allow adding an alarm with the lock already held.
It also required a bugfix for a latent bug in mem_lock_manager.cc
about capturing a temp StringPiece in the map when querying it via
operator[]; it must be queried using find().

In this wrapper we use the scheduler-mutex for protecting the
MemLockManager, but we ensure that we are not holding that mutex when
we call the NamedLock callbacks.  All of the code in this CL is around
doing that, and connecting the wakeups/alarms of the scheduler with
those needed by the MemLockManager.

Another commit will follow, which tests the migration of the system-test and
unit-test framework to use this lock manager.

[work by jmarantz; committed by jefftk]