ApacheFetch: add buffering
Before this change we would call ap_rwrite() etc from the rewrite thread, and if
that blocked for a while we might not have any rewrite threads available to
serve other requests.  With this change, all potentially blocking Apache calls
always happen on the request thread.  Writes are buffered in ApacheFetch until
the resource is complete, and then sent out in one go.

Most uses of this (ex: IPRO) were already not streaming, so we don't lose that
much by buffering.  We are doing more copying than we were, and to evaluate this
impact I ran "siege http://localhost:8080/$testimage -c200 -t1h" both with and
without the change, to really stress test this piece of ipro.  This test pulls a
single 1.1MB ipro optimized image with 200 concurrent readers for 1hr, and
should give us a worst-case indication of the slowdown buffering causes.

    before qps: 393.11
    after qps: 392.20

A 0.2% worstcase slowdown is not bad at all; buffering seems to not be a
problem.

This fixes some instances of #1048, but we're not sure yet whether it fixes all
of them.  It definitely doesn't fix ones due to slow filesystems, but we haven't
seen that version in the wild.

For uses of ApacheFetch that we know will always be synchronous we disable this
buffering, which is slightly more efficient.