fix apr_memcache multiget to prevent spinning
fixes #1048

in the case where
           if (strncmp(MS_VALUE, conn->buffer, MS_VALUE_LEN) == 0) {
and
           else if (strncmp(MS_END, conn->buffer, MS_END_LEN) == 0) {

both fail, it was possible for queries_sent to never decrement.
This patch sets rv to APR_EGENERAL in this case, decrements the queries_sent,
and closes the connection.

According to the trace from betabrand this is where the hang is.

Patch applied from apr dev mailing list
(http://www.mail-archive.com/dev%40apr.apache.org/msg26265.html)

Also adds a test for this path. The test is dependent on a new print to
stderr in apr_memcache2, but the real test is that we don't hang in this
situation.