Read in file-cache entries all at once, rather than a chunk at a time. When reading large images a chunk at a time, callgrind indicated a significant amount of time spent appending bytes to a dynamically expanding buffer.
Note: this adds a call to fstat so we know the size of the file before
we start to read it.

Another note: the only way I could find to implement this reading into
std::string involved using std::string::resize, which nulls the bytes we
are about to write.  That seems like a shame but it is what it is :).