Strip subresource hints / issue 973 Default behaviour is to strip subresource links which are in scope for pagespeed, these are the resources that are not disallowed or are valid domains in the domain laywer.
Added can_modify_url flag to HtmlParse and CanModifyUrl function to the HtmlFilter which
indicates whether urls can be rewritten by the parser and thus should be removed.
This is tested now in the strip_subresource_hints_filter_test.cc as this is only used by the
strip subresource hints feature right now. This should be moved to the HtmlParse tests.

DetermineEnabledFilters has been rolled up into DetermineFiltersBehaviour, which also
determines can_modify_url for all the filters and possible future "behaviors".

Added new option to explicitly prevent the default behaviour:
ModPreserveSubresourceHints on/off

For ubuntu a check for the new setup /var/www/html instead of /var/www for the document root
has been added.