Remove our dependency on apr_sockaddr_ip_getbuf
The new version of serf uses apr_sockaddr_ip_getbuf to add IP addresses to log
messages, but this function is new in APR 1.3.  This means we get a runtime
error on CentOS 5, where Apache fails to start.

We didn't catch this on our buildbot because one of our dependencies needs to be
fetched over svn with svn 1.8 or higher, and installing the new svn brought
along a new apr (1.5) and everything worked.

Fixes #1224