Set x-original-content-length for optimizations, including minification & gzip compression.
Maintain the semantic that when decompressing, if we lost all our reduction,
we remove the x-original-content-length.

If we have multiple levels of compression, we retain the largest value
of x-original-content-length.

Compute x-original-content-length across resource rewrites even if we don't
have a content-length header on the origin, but we do have the entire content
loaded.