Check that certificates are valid for the domain we're trying to contact.
This depends on x509_check_host, which was added in openssl 1.0.2.  If we want
to support older versions of openssl we need to parse certs ourself, like svn
does, which is a bunch of moderately tricky security sensitive code that I'd
really rather avoid.  We normally build agaist boringssl, which has this
function, but we also prepare a tarball build that intendes to link against
system openssl.  So anyone using that build process will need to upgrade to
1.0.2.