Optionally run mod_pagespeed ipro rewrite tasks on the request thread.
Change to having ApacheFetch block using the scheduler rather than
TimedWait on its own condvar, whenever there is a RewriteDriver.  This
allows in-place deadlines to work, and eliminates the 'abandoned' state
of ApacheFetch, simplifying its code and usage.

Added a Sequence* to CacheUrlAsyncFetcher in order to synchronize the
memcached cache response with the request-thread.