Add centralized process
This change adds two processes: a controller, and a babysitter. The
controller process is intended to host cheesy's controller work. The
babysitter process just waits around in case the controller dies and
restarts it if it does.

In order to exit when the host process exits, the controller watches a
pipe.  If reading from that pipe gets an EOF, then it knows the master
quit and it can quit itself. If the master wants to quit the controller
to load a new one with an updated configuration, it can write a byte to
the pipe, which the controller will see and exit.

Reviewed in https://github.com/pagespeed/mod_pagespeed/pull/1260

Companion nginx change: https://github.com/pagespeed/ngx_pagespeed/pull/1113