Fix request processing stalling on H/2 POSTS (as of nginx 1.9.14)
Nginx 1.9.14 introduces changes that seem to make us stall request
processing when we receive a POST request on HTTP/2. Debugging shows
we initiate the IPRO lookup, resume processing and things do not
proceed after that.

I suspect this is because NGINX now expects us to initiate reading of
the request body before we embark on an async wait operation in the
preaccess phase, but I yet have to prove that.

This change makes us only accept doing work for POST requests for
beacons in the preaccess phase. This path actually read the request
body and passes manual tests. I think is sufficient to tackle the
problem - and it looks like a change we should make anyway.

Fixes https://github.com/pagespeed/ngx_pagespeed/issues/1175