Make ngx_pagespeed behave like mod_pagespeed with regard to Flushing
This change makes ngx_pagespeed listen to the FollowFlushes option.
When set to on (=default), ngx_pagespeed will forward incoming flushes
to ProxyFetch. When writing output, we'll now also set the flush flag on
the buffers we are about to send downstream.

Companion to mps PR https://github.com/pagespeed/mod_pagespeed/pull/1066