Add InMemoryCache implementation and tests (partially in 74e2f95)
InMemoryCache is a maximally simple implementation of cacheInterface,
for debugging and faking purposes.

This changelist is more for me (Egor) to get feedback than to actually submit the
code into Piper, because I doubt there are any direct applications. The goal is to ensure that I understand what it's going on in caches and that I haven't missed anything important.