system-tests: give more information when fetch-until fails
When debugging system test flakes that manifest as fetch_until timeouts it's
helpful to have the failed input to fetch_until available in the failure
message, because that's what people put in bug reports (and is all that's
available on blaze).  This change makes fetch_until print:

  Fetched file: 681 bytes (
  ... contents of fetch file ...
  )

on timeouts.  If the fetched file is binary we just dump the first 256 characters, as hex, along with how big it is and its guessed mime type:

  Fetched file: 98051 bytes, image/jpeg; charset=binary (
  00000000: ffd8 ffe0 0010 4a46 4946 0001 0100 0001  ......JFIF......
  00000010: 0001 0000 ffdb 0043 0005 0304 0404 0305  .......C........
  00000020: 0404 0405 0505 0607 0c08 0707 0707 0f0b  ................
  [snip]
  000001f0: 0000 0000 0000 0000 0000 0000 0000 0000  ................
  ...
  )

Inspired by https://github.com/pagespeed/mod_pagespeed/issues/1359