caching: remove min_cache_ttl_ms option
We added an experimental option to extend the cache lifetime of
resources that explicitly set a TTL below our minimum.  It looks to me like this
can be retired now that the experiment is done.

The option was never exposed in open source, so it's safe to fully remove it.