cache: checkpoint the shm cache to disk
Snapshot the SHM cache to disk about every five minutes, and restore on restart.  Stop writing changes to the shm metadata cache through to the file cache.

Nginx side of the change, which includes a system test: https://github.com/pagespeed/ngx_pagespeed/pull/930

(No system test for Apache because what we're testing is completely shared between the two of them and we don't already have restart tests for Apache.)