Refactor the way the sed replacements are constructed for hacking debug.conf.template into pagespeed.conf, using gmake functions.  Add an ending delimiter to all the sed-patterns for customizing the configuration templates.
The refactoring reduces slightly the number of times we must repeat
stanzas involving OPT_* et al in our makefiles.  More reductions are
needed.

The ending delimeter (\b) solves a problem where the pattern "#REDIS"
was a substring of "#REDIS_LOADTEST", and sed would break the template.
To work around this problem I used "#RED_LOADTEST" but now you don't
find that variable with a csearch for "REDIS".