Suppress "Adding function to sequence after shutdown" message on release builds.
I noticed this when testing the controller which can wake up a bunch of
blocked rewrites as part of the shutdown process. However, in general these
messages are generally useless/worrying for end users, plus we DTRT by
canceling the Function cleanly anyway. The gRPC specific issue will be
discussed when that code lands.