Support building against latest-beta.
This is somewhat awkward, since we need to know what the url is for the matching PSOL binary.  I'm adding a PSOL_BINARY_URL file to the repo that tells us where to download it from, but for old versions we have to grep it out of the config file.

I also rewrote the error message you get in the config file telling you how to donwload PSOL.  Now it knows whether you're working with a release or a development branch, and tells you either where to download psol or how to build it from from source