Fix a ~5% flake in AprMemCacheTest when run at Google.  It might be due to signals being sent & caught, and a slightly-too-tight timeout.
The root cause of the interrupt problem was lack of EINTR handling in
apr_memcache2.c.  In general in Apache there is no signal handling so
this does not affect any kind of production use (at least when not
shutting down), but affects our unit tests in some enviornments.  As far
as I can tell there was only one point where the EINTR handling needed
to be added, but I left behind more debug printing, via
fprintf(stderr...), which came through in unit tests.  No
message-handlers are available from inside apr_memcache2.c.

Backs out the change to send an ascii reset sequence to the memcache.
It's totally unnecessary, exposes the ascii protocol of memcached at
the wrong layer of abstraction, and adds further risk to the test.