debian-install: put pagespeed-libraries.conf where it will be loaded
The way you used to install apache2 configuration on Debian-based systems
was to put it in /etc/apache2/conf.d/, but starting with Ubuntu 14 LTS
there's a new system where you install it to /etc/apache2/conf-available/ and
then run a2enconf to symlink it to /etc/apache2/conf-enabled.  On these
systems we weren't installing pagespeed-libraries.conf in a way that made
it usable, because when we put it in conf.d it just was never loaded.

Switch to installing to /etc/apache2/conf-available/, and then use a2enconf
if it's available.  If a2enconf isn't available, manually symlink to conf.d.

Fixes: https://github.com/pagespeed/mod_pagespeed/issues/1389