system tests: remove running tests in the background
This was added, found to be too flaky, and has been disabled for over a year.  It probably doesn't work anymore, and if we want to revive it we'll probably do it a different way.

(Also, I want to make changes to this part of the test script, and having stale code around makes it hard to make sure I haven't broken the stale code.)