Fix handling of failed AsyncFetches.
AsyncFetch::HandleDone(success=false) can't pass success=false to
HeadersComplete.  Success being false means "we've encountered an error"
but the headers may or may not have been updated to match.  For example,
if the CacheUrlAsyncFetcher doesn't find something it will set
status=false and set up 404 headers, but if SendFallbackResponse aborts
a response for having the wrong content type then it sets status=false
while leaving 200 headers.

We need to make sure that error headers are sent whenever status=false,
and the only place we can make sure of that is in AsyncFetch::Done
itself, because HandleDone is too late and HandleHeadersComplete can't
see status.

This fixes a flake in resource_content_type_html.sh with ngx_pagespeed
where when the deadline alarm was hit, which mostly happens under
valgrind, it would send out the headers for the request with no body.

I've verified the fix by making sure that the test passes if you have
set test_instant_fetch_rewrite_deadline=true, and also that while we
previously got 5 flakes in 133 runs under valgrind now we don't get any.

(I previously tried to fix this on the ngx_pagespeed end with
pagespeed/ngx_pagespeed#1307 , but that doesn't work consistently
because ngx_pagespeed races to send out headers immediately after
receiving HeadersComplete.)