test: run_tests.sh should just assume 8050 and 8051
Right now run_tests.sh assumes it has 8053 and 9991 without causing any
problems.  Start assuming we have 8050 and 8051 as well instead of
requiring people to pass them in.  If this becomes a problem we can add
a way to override ports without editing the config.

Also, switch travis to actually run the tests instead of just echoing
the command to run them.