Fix race that could sometimes cause us to not dispatch some background fetches:
1) RateController::Fetch would try IncrementIfCanTriggerFetch, and fail since there were too many outstanding background fetches for the host. Note that "too many" is 1 in case of prefork.
2) A previous fetch would finish, notice there is nothing queued, so do nothing else.
3) RateController::Fetch would EnqueueFetchIfWithinThreshold, but there is no long and outstanding fetch for the host to push it out of the queue on completion.

The change just locks the HostFetchInfo for the duration, and adds a bunch of lock annotation for that. (This does mean we hold two locks now).

(This was probably more visible in loadtests since this is per-host and that has lots
 of hosts).

Also fix an occasional DCHECK about current_global_fetch_queue_size_, where we could end up dequeing the entry and decrementing it before the increment due to another race --- thanks to Josh for pointing it out.